Update CLI tool to use ESLint 8 (CLIEngine class is gone)
Bumped minimum engine as it now uses top-level await.
