+++ packages/check-es-compat/package.json
@@ -27,13 +27,13 @@
   "license": "MIT",
   "author": "Robat Williams",
   "bin": {
-    "check-es-compat": "bin/cli.js"
+    "check-es-compat": "bin/cli.mjs"
   },
   "dependencies": {
-    "eslint": "^6.7.2",
+    "eslint": "^8.6.0",
     "eslint-plugin-ecmascript-compat": "^1.1.1"
   },
   "engines": {
-    "node": ">=14.0.0"
+    "node": ">=14.8.0"
   }
 }
